requirements.txt
python-dateutil
requests
requests_oauthlib
jira
PyYAML